Isolate performance diagnostics projection
